==> code/hm_notebooks/2_merging_smoke_fires.ipynb <==
import pandas as pd
import numpy as np
import geocoder

path = '../scratch_data/'
smoke = pd.read_csv(path + 'smoke_west_counties_2010_2016.csv')

print(smoke.shape)
smoke.head()

fires = pd.read_csv(path + 'mtbs_FODpoints_DD/MTBS_FODpoints.csv', lineterminator='\n')
wildfires = fires.copy()

print(fires.shape)
fires.head()

smoke.columns = smoke.columns.str.lower()
wildfires.columns = wildfires.columns.str.lower()

#drops - only run cell once
wildfires.drop('unnamed: 0', axis=1, inplace=True)
wildfires['geometry2'] = wildfires['geometry'].map(lambda x: x.replace('POINT (', ''))

### Try merge on lat/long
* wildfires: extract lat/lon from 'geometry'
    * same as 'brnbndlat' and 'brnbndlon' with more decimals...
    * filter df by lon (since no 'state' id) : min_lon_west = -125.5, max_lon_east = -108.5
* smoke: create lat/lon from county

wildfires[['lon', 'lat']] = wildfires['geometry2'].str.split(' ', expand=True)
wildfires['lat'] = wildfires['lat'].str.replace(')','').astype('float')
wildfires['lon'] = wildfires['lon'].astype('float')

wildfires['ig_date'] = wildfires['ig_date'].str.replace('-','').astype('int32')

wildfires = wildfires.loc[(wildfires['ig_date']>=20100604) & (wildfires['ig_date']<=20161231)]

### lat/lon too granular for merge easily with smoke, try reverse geocode county to wildfires

len(wildfires)

#bound by lat/lon to get States of interest, no 'state' identifier
wildfires = wildfires.loc[wildfires['lon'] <= -108.5]  #eastern bound, AZ-NM border
wildfires = wildfires.loc[wildfires['lat'] <= 53.5] #northernmost point in smoke data
wildfires = wildfires.loc[wildfires['lon'] >= -130] #west bound since apparently HI is in here

len(wildfires)

wildfires.head(1)

reverse_geo_lats = wildfires['lat'].to_list()
reverse_geo_lon = wildfires['lon'].to_list()
rev_geo_coords = list(zip(reverse_geo_lats, reverse_geo_lon))

min(reverse_geo_lats), max(reverse_geo_lats)

min(reverse_geo_lon), max(reverse_geo_lon)

from geopy.geocoders import Nominatim
geolocator = Nominatim(user_agent = 'helen.meigs@gmail.com')
location = geolocator.reverse(rev_geo_coords[0])

location[0]

# %%time
# wf_counties = []
# wf_states = []
# wf_full_loc = []

# for i in range(len(rev_geo_coords)):
#     location = geolocator.reverse(rev_geo_coords[i])
#     county = location[0].split(',')[0]
#     state = location[0].split(',')[1]
#     wf_counties.append(county)
#     wf_states.append(state)
#     wf_full_loc.append(location[0])

##### Tried working with geocoded 'location' output in list  form (each entry a single str that could be split on ',') but couldn't figure it out.  Proceed to the less elegant answer:
* New df with geo_columns for each section of location
    * _ONE_ of these geo_columns will contain 'County' for most observations
    * geo_column containing 'County' will vary
        * curve ball, some things like County Road, and 'Not in a County' exist. great.
* Create a column for _each?_ geo_col indicating presence/absence of 'County'?
* Populate some other final column with all the counties?  Maybe by cascading fillna()?
* REPEAT WITH STATE

# wf_locdf = pd.DataFrame(wf_full_loc, columns=['reverse_geocode'])
# wf_locdf.to_csv('../scratch_data/mtbs_FODpoints_DD/mtbs_reverse_geocode.csv', index=False)

wf_locdf = pd.read_csv('../scratch_data/mtbs_FODpoints_DD/mtbs_reverse_geocode.csv')

wf_locdf.head()

wf_locdf[['geo1', 'geo2', 'geo3', 'geo4', 'geo5', 'geo6', 'geo7']] = wf_locdf['reverse_geocode'].str.split(',', expand=True)

wf_locdf.fillna('', inplace=True)

geo_cols = wf_locdf.drop(columns='reverse_geocode').columns.to_list()

for i in geo_cols:
    wf_locdf[f'{i}_county'] = wf_locdf[i].map(lambda x: x if 'County' in x else np.nan)

wf_locdf['county_name'] = np.nan

county_cols = ['geo1_county', 'geo2_county', 'geo3_county', 'geo4_county', 'geo5_county']
for i in county_cols:
    wf_locdf['county_name'].fillna(wf_locdf[i], inplace=True)

# only 15 locations without county ID, not bad
wf_locdf.info()

##### Repeat for State, so can filter by State
More annoying because States don't say the word State in them...luckily I only need seven.

wf_locdf.head()

# westofrockies = ['Montana', 'Wyoming', 'New Mexico', 'Colorado', 'California', 'Arizona', 'Nevada', 'Washington', 'Oregon', 'Utah', 'Idaho']
# for i in geo_cols:
#     wf_locdf[f'{i}_state'] = wf_locdf[i].str.contains('|'.join(westofrockies)).astype('int')

# for i in geo_cols:
#     for j in range(len(wf_locdf)):
#                    if wf_locdf[f'{i}_state'][j] == 1:
#                        wf_locdf[f'{i}_state'][j]=wf_locdf[i][j]
#                    else:
#                        wf_locdf[f'{i}_state'][j]=np.nan
            
        

# wf_locdf['state'] = np.nan
# state_cols = ['geo1_state', 'geo2_state', 'geo3_state', 'geo4_state', 'geo5_state', 'geo6_state']
# for i in state_cols:
#     wf_locdf['state'].fillna(wf_locdf[i], inplace=True)

# wf_locdf.drop(columns=geo_cols + county_cols + state_cols, inplace=True)
# wf_locdf.drop(columns = ['geo6_county', 'geo7_county', 'geo7_state'], inplace=True)

# pd.set_option('display.max_colwidth', 1000)
# wf_locdf.loc[wf_locdf['county_name'].isnull()]  #now what

# #manually impute county nulls by google mapping it and seeing what's closest
# wf_locdf['county_name'][94] = 'Duchesne County'
# wf_locdf['county_name'][126] = 'Santa Cruz County'
# wf_locdf['state'][126] = 'Arizona'
# # index 135 cannot impute county (Utah, U.S.)
# wf_locdf['county_name'][188] = 'San Diego County'
# wf_locdf['state'][188] = 'California'
# wf_locdf['county_name'][288] = 'Los Angeles County'
# wf_locdf['county_name'][323] = 'Duchesne County'
# wf_locdf['county_name'][486] = 'Cochise County'
# wf_locdf['state'][486] = 'Arizona'
# wf_locdf['county_name'][589] = 'La Paz County'
# # index 649 cannot impute county (Utah, U.S.)
# wf_locdf['county_name'][698] = 'Los Angeles County'
# wf_locdf['county_name'][836] = 'Los Angeles County'
# wf_locdf['county_name'][1045] = 'Weber County'
# wf_locdf['county_name'][1055] = 'Los Angeles County'
# wf_locdf['county_name'][1177] = 'Los Angeles County'
# wf_locdf['county_name'][1500] = 'Los Angeles County'
# wf_locdf['county_name'][1504] = 'Los Angeles County'
# # index 1528 cannot impute county (Utah, U.S.)

# wf_locdf.loc[wf_locdf['county_name'].isnull()] #checks out

Got a bunch of not-States, dangit (ie Washington Camp, which is in Arizona).  Few, so fix manually.

# #got a bunch of not-states, dangit
# wf_locdf.state.value_counts()

# wf_locdf.loc[wf_locdf['state'].str.contains('County')] #phew easy fix

# wf_locdf['state'] = wf_locdf['state'].str.replace('County','')

# wf_locdf['state'] = wf_locdf['state'].str.replace('Riding and Hiking ','')
# wf_locdf['state'] = wf_locdf['state'].str.replace('Trail','')
# wf_locdf['state'] = wf_locdf['state'].str.replace('Washingtonia Drive','California') #San diego county
# wf_locdf['state'] = wf_locdf['state'].str.replace('Washington Camp','Arizona') #in AZ

# #strip leading and trailing whitespace for state and county_name
# wf_locdf['state'] = wf_locdf['state'].str.strip()
# wf_locdf['county_name'] = wf_locdf['county_name'].str.strip()

# wf_locdf.to_csv('../scratch_data/mtbs_FODpoints_DD/mtbs_reverse_geocode.csv', index=False)

## Merge reverse geocodes back to wildfires
Now wildfire df has state and county name
* Few NANs
* few obs from States not of interest (Wyoming, Colorado, New Mexico, Real Mexico).  However, these made the cut based on the lat/lon boundaries established earlier so these fire ignition points are very close to our established bounds (close to state border).
    * keep in air quality analysis, because they will impact pollution sensors even though fires not originating from 'State of interest'?

wf_locdf = pd.read_csv('../scratch_data/mtbs_FODpoints_DD/mtbs_reverse_geocode.csv')

wildfires.reset_index(drop=True, inplace=True)

wildfires_with_counties = pd.concat([wildfires, wf_locdf], axis=1)

wildfires_with_counties.head()

wildfires_with_counties.columns

wildfires_with_counties.drop(['irwinid', 'dnbr_offst', 'dnbr_stddv', 'pre_id', 'post_id',
                'map_prog', 'map_id', 'comment', 'geometry2'], axis=1, inplace=True)

wildfires_with_counties.rename(columns = {'ig_date':'date'}, inplace=True)

wildfires_with_counties.info()

#manually impute missing counties based on incident name & lat/lon score
wildfires_with_counties['county_name'][135] = 'Uintah County'
wildfires_with_counties['county_name'][649] = 'Uintah County'
wildfires_with_counties['county_name'][1528] = 'Uintah County'

## Merge wildfires df with smoke df
## EDIT: first add lat/lon to smoke before merging
Analyzing at county level or higher, but want lat/lon for every observation for mapping or potential clustering (clusters may be better than counties to group nearby observations)

### Adding lat/lon to smoke - geocodes from county_name
Code to generate county_key is at end of notebook (had sidelined because thought no longer useful).  Starting with completed county_key here:

Geocode based on county name AND STATE NAME.  For example, there is a Kings County CA and Kings County PA, so your first go at this did not pan out and you never even knew it until it was mapped out in Tableau and you had smoke measurements coming from the East coast.

## Smoke df: Need to make column with 'county_name, State' for geocoding
Otherwise you get counties from East Coast with shared names

smoke.state.value_counts() #yes filtered to what we want

smoke['county_state'] = smoke['county_name'] + ', ' + smoke['state']

smoke.head()

unique_locs = set(smoke.county_state)

county_key = pd.DataFrame(unique_locs, columns=['county_state'])

county_key.loc[county_key['county_state'].str.contains('King')]

# from geopy.geocoders import Nominatim
# geolocator = Nominatim(user_agent = 'helen.meigs@gmail.com')

# %%time
# county_key['lat']= county_key['county_state'].map(lambda x: geocoder.arcgis(x).latlng[0])
# county_key['lon'] = county_key['county_state'].map(lambda x: geocoder.arcgis(x).latlng[1])

# # wall time 6min

# county_key['county_name'] = county_key['county_state'].map(lambda x: x[:-4])

# county_key.to_csv('../scratch_data/smoke_2000_16_county_key.csv', index=False)

county_key = pd.read_csv('../scratch_data/smoke_2000_16_county_key.csv')

county_key.head()

smoke = pd.merge(smoke, county_key, on=['county_state', 'county_name'])

smoke.rename(columns={'lat':'lat_smo', 'lon':'lon_smo'}, inplace=True)

smoke.drop(columns = ['statefp', 'countyfp', 'population','county_state'], axis=1, inplace=True)
smoke.drop_duplicates(inplace=True)

smoke.reset_index(drop=True, inplace=True)

smoke.head()

## Back to merging wildfires with smoke df:

wildfires_with_counties.head()

def rename_latlon(df, name):
    df.rename(columns = {'lat':f'lat_{name}', 'lon':f'lon_{name}'}, inplace=True)
    
    
rename_latlon(smoke, 'smo')
rename_latlon(wildfires_with_counties, 'wf')

state_dict = {'Idaho':'ID', 'California': 'CA', 'Arizona':'AZ', 'Oregon':'OR', 
              'Washington':'WA', 'Nevada':'NV', 'Montana':'MT', 'Utah':'UT',
              'Wyoming':'WY', 'New Mexico':'NM', 'Colorado':'CO'}

wildfires_with_counties['state_abbrev'] = wildfires_with_counties['state'].map(state_dict)

wildfires_with_counties.rename(columns = {'state':'state_full', 'state_abbrev':'state',
                                         'reverse_geocode':'reverse_geocode_fire_ig'}, inplace=True)

wildfires_with_counties.tail(1)

smoke.tail(1)

inner_fires_smoke = pd.merge(smoke, wildfires_with_counties, on = ['county_name','state', 'date'], how='inner')
keepfires_smoke = pd.merge(smoke, wildfires_with_counties, on = ['county_name', 'state', 'date'], how='right')
outer_fires_smoke = pd.merge(smoke, wildfires_with_counties, on=['county_name', 'state', 'date'], how = 'outer')

See len of each join: inner join drops 100 fires (of 1588 - no good).  Maintaining all fires gives 2250 datapoints, but we do want pollution data for days in between fires, so the outer join of 83,000 may be what we have to work with. Especially because this is only fire IGNITION DATES, and we absolutely want to look at days immediately post-fire, and how they compare to baseline (long duration away from fire).  Exporting outer join for EDA, the limited versions are option for modeling.

print(len(smoke))
print(len(wildfires_with_counties))
print(len(smoke) + len(wildfires_with_counties))
print(len(outer_fires_smoke), len(keepfires_smoke), len(inner_fires_smoke))

inner_fires_smoke

outer_fires_smoke.drop_duplicates(inplace=True)
outer_fires_smoke.reset_index(drop=True, inplace=True)

#outer_fires_smoke.to_csv('../scratch_data/MTBS_merge_smoke_inProgress.csv', index=False)
