==> code/hm_notebooks/4_annual_pdfs_to_csv.ipynb <==
import numpy as np
import pandas as pd
import tabula
import os

pd.set_option('display.max_rows',None)

# Annual fire #s and acres - wild and Rx - by State
Methods for cleaning determined in separate shitshow of a notebook, final functions here.

* All States of interest contained on pages 1, 2, 3, 7, 8, 10 of [NICC Redbooks](https://www.nifc.gov/fire-information/statistics) for years 2010-2013.
    * p1, p2-8, and p10 all require separate treatment because tabula likes to make you work for it
    * function for each page defined separately before combining, because possible that the pdfs per year are _slightly_ different enough to break the combined function; if have to resort to processing individual documents manually, easier to do with functions broken out

* Format slightly changes in 2014; new function required.

west_states = ['CA', 'OR', 'WA', 'AZ', 'NV', 'UT', 'ID']

### p1 methods

#define function for p1 cleaning.  
#all other pages way worse, but in the same way so can clean together

def p1_clean(filename):
    west_states = ['CA', 'OR', 'WA', 'AZ', 'NV', 'UT', 'ID']
    table = tabula.read_pdf(filename, pages=1)
    
    df = table[0].copy()
    df.rename(columns = {'Unnamed: 0': 'state', 'Unnamed: 1':'agency', 'Wildland':'fires_acres', 'Unnamed: 2': 'rx_fires', 'Rx': 'rx_acres'}, inplace=True)
    df.drop(index=0, inplace=True)
    df[['wildfires', 'wild_acres']] = df['fires_acres'].str.split(expand=True)
    df.drop('fires_acres', axis=1, inplace=True)
    df['state'] = df['state'].fillna(method='ffill')
    df = df.loc[df['state'].isin(west_states)]
    
    return df

### p2-8 methods

# define function for p2-9 cleaning

def p2to9_clean(filename):
    west_states = ['CA', 'OR', 'WA', 'AZ', 'NV', 'UT', 'ID']
    table = tabula.read_pdf(filename, pages=1)
    
    df = table[0].copy()
    
    df['Wildland'] = df['Wildland'].fillna('Totals')
    df.drop(columns=['Unnamed: 1', 'Unnamed: 2', 'Unnamed: 10'], inplace=True)
    
    df['Rx'].fillna(df['Unnamed: 3'], inplace=True)
    df['Unnamed: 4'].fillna(df['Unnamed: 5'], inplace=True)
    df['Unnamed: 6'].fillna(df['Unnamed: 7'], inplace=True)
    df['Unnamed: 8'].fillna(df['Unnamed: 9'], inplace=True)
    
    df.drop(columns=['Unnamed: 3', 'Unnamed: 5', 'Unnamed: 7', 'Unnamed: 9'], inplace=True)
    
    df.rename(columns={'Unnamed: 0':'state', 'Wildland': 'agency', 'Rx':'wildfires',
                     'Unnamed: 4':'wild_acres', 'Unnamed: 6':'rx_fires', 
                      'Unnamed: 8': 'rx_acres'}, inplace=True)
    
    df.drop(index=0, inplace=True)
    
    df['state'] = df['state'].fillna(method='ffill')
    
    df = df.loc[df['state'].isin(west_states)]
    
    return df

# define function for p10 cleaning: ok very few changes from p2 I overreacted

def p10_clean(filename):
    west_states = ['CA', 'OR', 'WA', 'AZ', 'NV', 'UT', 'ID']
    table = tabula.read_pdf(filename, pages=1)
    
    df = table[0].copy()
    
    df['Unnamed: 1'] = df['Unnamed: 1'].fillna(df['Rx'])
    df.drop(columns=['Rx', 'Wildland', 'Unnamed: 10'], inplace=True)
    
    df['Unnamed: 2'].fillna(df['Unnamed: 3'], inplace=True)
    df['Unnamed: 4'].fillna(df['Unnamed: 5'], inplace=True)
    df['Unnamed: 6'].fillna(df['Unnamed: 7'], inplace=True)
    df['Unnamed: 8'].fillna(df['Unnamed: 9'], inplace=True)
    
    df.drop(columns=['Unnamed: 3', 'Unnamed: 5', 'Unnamed: 7', 'Unnamed: 9'], inplace=True)
    
    df.rename(columns={'Unnamed: 0':'state', 'Unnamed: 1': 'agency', 'Unnamed: 2':'wildfires',
                     'Unnamed: 4':'wild_acres', 'Unnamed: 6':'rx_fires', 
                      'Unnamed: 8': 'rx_acres'}, inplace=True)
    
    df.drop(index=0, inplace=True)
    
    df['state'] = df['state'].fillna(method='ffill')
    
    df = df.loc[df['state'].isin(west_states)]
    
    return df

### United into big ol function

# from scratch (all functions in here instead of defined elsewhere)

def pdf_no_more(input_path, output_path, year):

    pdf_list = os.listdir(input_path)
    
    west_states = ['CA', 'OR', 'WA', 'AZ', 'NV', 'UT', 'ID']
    
    df_list = []
    
    for i in pdf_list:
        filepath = input_path + i

        table = tabula.read_pdf(filepath, pages=1)
        df = table[0].copy()

        
        if int(filepath[-5]) == 1:
            print(f'processing page Alpha {filepath[-5]}')

            df.rename(columns = {'Unnamed: 0': 'state', 'Unnamed: 1':'agency', 'Wildland':'fires_acres', 'Unnamed: 2': 'rx_fires', 'Rx': 'rx_acres'}, inplace=True)
            df.drop(index=0, inplace=True)
            df[['wildfires', 'wild_acres']] = df['fires_acres'].str.split(expand=True)
            df.drop('fires_acres', axis=1, inplace=True)
            df['state'] = df['state'].fillna(method='ffill')
            df = df.loc[df['state'].isin(west_states)]
            
            df_list.append(df)

            
        if int(filepath[-5]) == 0: #page 10 will read as 0 - only last digit referenced
            
            print(f'processing page Omega {filepath[-5]}')
            df['Unnamed: 1'] = df['Unnamed: 1'].fillna('Totals')
            df.drop(columns=['Rx', 'Wildland', 'Unnamed: 10'], inplace=True)

            df['Unnamed: 2'].fillna(df['Unnamed: 3'], inplace=True)
            df['Unnamed: 4'].fillna(df['Unnamed: 5'], inplace=True)
            df['Unnamed: 6'].fillna(df['Unnamed: 7'], inplace=True)
            df['Unnamed: 8'].fillna(df['Unnamed: 9'], inplace=True)

            df.drop(columns=['Unnamed: 3', 'Unnamed: 5', 'Unnamed: 7', 'Unnamed: 9'], inplace=True)

            df.rename(columns={'Unnamed: 0':'state', 'Unnamed: 1': 'agency', 'Unnamed: 2':'wildfires',
                             'Unnamed: 4':'wild_acres', 'Unnamed: 6':'rx_fires', 
                              'Unnamed: 8': 'rx_acres'}, inplace=True)

            df.drop(index=0, inplace=True)
            df['state'] = df['state'].fillna(method='ffill')
            df = df.loc[df['state'].isin(west_states)]
            
            df_list.append(df)
        
        if 1 < int(filepath[-5]) & int(filepath[-5]) < 10:

            print(f'processing page Medio {filepath[-5]}')

            df['Wildland'] = df['Wildland'].fillna('Totals')
            df.drop(columns=['Unnamed: 1', 'Unnamed: 2', 'Unnamed: 10'], inplace=True)

            df['Rx'].fillna(df['Unnamed: 3'], inplace=True)
            df['Unnamed: 4'].fillna(df['Unnamed: 5'], inplace=True)
            df['Unnamed: 6'].fillna(df['Unnamed: 7'], inplace=True)
            df['Unnamed: 8'].fillna(df['Unnamed: 9'], inplace=True)

            df.drop(columns=['Unnamed: 3', 'Unnamed: 5', 'Unnamed: 7', 'Unnamed: 9'], inplace=True)

            df.rename(columns={'Unnamed: 0':'state', 'Wildland': 'agency', 'Rx':'wildfires',
                             'Unnamed: 4':'wild_acres', 'Unnamed: 6':'rx_fires', 
                              'Unnamed: 8': 'rx_acres'}, inplace=True)

            df.drop(index=0, inplace=True)
            df['state'] = df['state'].fillna(method='ffill')
            df = df.loc[df['state'].isin(west_states)]
            
            df_list.append(df)
            
    final = pd.concat(df_list)
#     final.to_csv(f'{output_path}/annual_{year}_agency.csv', index=False)
    
#     final.loc[final['agency']=='Totals'].to_csv(f'{output_path}/annual_{year}_tot.csv', index=False)
    
    return final

input_path_1 = '../scratch_data/pdf_extracts/bystate_2010_extracted-pages/'
input_path_2 = '../scratch_data/pdf_extracts/bystate_2011_extracted-pages/'
input_path_3 = '../scratch_data/pdf_extracts/bystate_2012_extracted-pages/'
input_path_4 = '../scratch_data/pdf_extracts/bystate_2013_extracted-pages/'
output_path = '../scratch_data/pdf_extracts/resulting_csvs/'

#pdf_no_more(input_path_4, output_path, 2013) #breaks on 2013.  we had a good run.

## Construct 2013 manually
#### Another big format change coming 2014-2016, so not worth automating

AZ_2013 = p1_clean('../scratch_data/pdf_extracts/bystate_2013_extracted-pages/2013-1.pdf')
AZ_2013 #checked against doc

def midpages_2013(filename):

    west_states = ['CA', 'OR', 'WA', 'AZ', 'NV', 'UT', 'ID']
    table = tabula.read_pdf(filename, pages=1)
    
    df = table[0].copy()
    
    df.drop(columns=['Unnamed: 1', 'Unnamed: 2', 'Unnamed: 3', 'Unnamed: 5', 'Unnamed: 7'], inplace=True)
    
    df.rename(columns={'Unnamed: 0':'state', 'Wildland': 'agency', 'Rx':'wildfires',
                     'Unnamed: 4':'wild_acres', 'Unnamed: 6':'rx_fires', 
                      'Unnamed: 8': 'rx_acres'}, inplace=True)
    
    df.drop(index=0, inplace=True)
    
    df['state'] = df['state'].fillna(method='ffill')
    
    df = df.loc[df['state'].isin(west_states)]
    
    return df

path = '../scratch_data/pdf_extracts/bystate_2013_extracted-pages/'
pdf_list = os.listdir(path)

all_paths = [path + i for i in pdf_list] #order = 3, 2, 10, 1, 7, 8

p2 = midpages_2013(all_paths[1])
p3 = midpages_2013(all_paths[0])
#p7 = midpages_2013(all_paths[4])  #no states of interest on this page
p8 = midpages_2013('../scratch_data/pdf_extracts/bystate_2013_extracted-pages/2013-8.pdf')  #broken ^^

pdf = '../scratch_data/pdf_extracts/bystate_2013_extracted-pages/2013-10.pdf'
p10 = tabula.read_pdf(pdf, pages=1)
p10 = p10[0]
#p10

p10.drop(columns=['Wildland', 'Rx', 'Unnamed: 3', 'Unnamed: 5', 'Unnamed: 7'], inplace=True)

p10.rename(columns={'Unnamed: 0':'state', 'Unnamed: 1': 'agency', 'Unnamed: 2':'wildfires',
                 'Unnamed: 4':'wild_acres', 'Unnamed: 6':'rx_fires', 
                  'Unnamed: 8': 'rx_acres'}, inplace=True)

p10.drop(index=0, inplace=True)

p10['state'] = p10['state'].fillna(method='ffill')

p10 = p10.loc[p10['state'].isin(west_states)]

p10

#oh good, this one goes to 11
pdf = '../scratch_data/pdf_extracts/bystate_2013_extracted-pages/2013-11.pdf'
p11 = tabula.read_pdf(pdf, pages=1)
p11 = p11[0]  #I thiiink same as mediopages function?

p11_final = midpages_2013('../scratch_data/pdf_extracts/bystate_2013_extracted-pages/2013-11.pdf')
p11_final

all_2013 = pd.concat([AZ_2013, p2, p3, p8, p10, p11_final])
all_2013.head()

all_2013.to_csv(output_path + 'annual_2013_agency.csv', index=False)
all_2013.loc[all_2013['agency'] == 'Totals'].to_csv(output_path + 'annual_2013.tot.csv', index=False)

## 2014-2015 completely different formats than priors
Curses.  All pages unique.  Proceeding manually for all, last push.

filename14 = '../scratch_data/pdf_extracts/2014.pdf'
filename15 = '../scratch_data/pdf_extracts/2015.pdf'

def clean1_newformat(filename):
    west_states = ['CA', 'OR', 'WA', 'AZ', 'NV', 'UT', 'ID']
    table = tabula.read_pdf(filename, pages=1)
    df = table[0].copy()
    df.drop(columns=['Unnamed: 2', 'Unnamed: 4', 'Wildland', 
                     'Unnamed: 5', 'Unnamed: 8', 'Rx'], inplace=True)
    df.rename(columns = {'Unnamed: 0':'state', 'Unnamed: 1':'agency', 'Unnamed: 3':'wildfires',
                        'Unnamed: 6': 'wild_acres', 'Unnamed: 7': 'rx_fires', 
                         'Unnamed: 9': 'rx_acres'}, inplace=True)
    df['agency'] = df['agency'].fillna('Totals')
    df['state'] = df['state'].fillna(method = 'ffill')
    df = df.loc[df['state'].isin(west_states)]

    return df

p1_2014 = clean1_newformat(filename14)

def hopefully_more_than_one_page(file, page): #nope just one page.
    west_states = ['CA', 'OR', 'WA', 'AZ', 'NV', 'UT', 'ID']
    table = tabula.read_pdf(file, pages=page)
    df = table[0].copy()
    
    df['Unnamed: 2'] = np.where(df['Unnamed: 2']=='Totals', df['Unnamed: 11'], df['Unnamed: 2'])
    
    #manually fix rx_fires_total outside of function for P2 ONLY
    
    df.rename(columns = {'Unnamed: 0': 'state', 'Wildland':'agency', 'Rx':'wildfires',
                       'Unnamed: 1':'wild_acres', 'Unnamed: 2':'rx_fires',
                       'Unnamed: 3': 'rx_acres'}, inplace=True)
    
    df['wild_acres'].fillna(df['Unnamed: 8'], inplace=True)
    df['wildfires'].fillna(df['Unnamed: 5'], inplace=True)
    df['rx_acres'] = df.rx_acres.fillna(df['Unnamed: 14'])
    
    df['state'].fillna(method='ffill', inplace=True)
    df['agency'] = [i if i not in west_states else 'Totals' for i in df['agency']]
    
    df.drop(df.filter(regex='Unnamed').columns, axis=1, inplace=True)
    df.drop(index=0, inplace=True)
    df = df.loc[df['state'].isin(west_states)]
    
        
    return df

p2_2014 = hopefully_more_than_one_page(filename14, 2) 
p2_2014['rx_fires'][9] = 515 #note to manually fix units in rx_fires_tot

table = tabula.read_pdf(filename14, pages=3)
p3 = table[0]

#now that you've started cut-pasting intsead of generalizing a function,
#this chunk is good for 3 pages
p3['Unnamed: 2'] = np.where(p3['Unnamed: 2']=='Totals', p3['Unnamed: 11'], p3['Unnamed: 2'])
p3.rename(columns = {'Unnamed: 0': 'state', 'Unnamed: 1':'agency', 'Wildland':'wildfires',
                   'Rx':'wild_acres', 'Unnamed: 2':'rx_fires',
                   'Unnamed: 3': 'rx_acres'}, inplace=True)
p3['wild_acres'].fillna(p3['Unnamed: 8'], inplace=True)
p3['wildfires'].fillna(p3['Unnamed: 5'], inplace=True)
p3['rx_acres'] = p3.rx_acres.fillna(p3['Unnamed: 14'])
p3['state'].fillna(method='ffill', inplace=True)
p3['agency'] = [i if i not in west_states else 'Totals' for i in p3['agency']]
p3.drop(p3.filter(regex='Unnamed').columns, axis=1, inplace=True)
p3.drop(index=0, inplace=True)
p3 = p3.loc[p3['state'].isin(west_states)]

table = tabula.read_pdf(filename14, pages=7)
p7 = table[0]

p7['Unnamed: 2'] = np.where(p7['Unnamed: 2']=='Totals', p7['Unnamed: 11'], p7['Unnamed: 2'])
p7.rename(columns = {'Unnamed: 0': 'state', 'Unnamed: 1':'agency', 'Wildland':'wildfires',
                   'Rx':'wild_acres', 'Unnamed: 2':'rx_fires',
                   'Unnamed: 3': 'rx_acres'}, inplace=True)
p7['wild_acres'].fillna(p7['Unnamed: 8'], inplace=True)
p7['wildfires'].fillna(p7['Unnamed: 5'], inplace=True)
p7['rx_acres'] = p7.rx_acres.fillna(p7['Unnamed: 14'])
p7['state'].fillna(method='ffill', inplace=True)
p7['agency'] = [i if i not in west_states else 'Totals' for i in p7['agency']]
p7.drop(p7.filter(regex='Unnamed').columns, axis=1, inplace=True)
p7.drop(index=0, inplace=True)
p7 = p7.loc[p7['state'].isin(west_states)]

p7['rx_fires'][9] = 27
p7['rx_fires'][36] = 658

table = tabula.read_pdf(filename14, pages=9)
p9 = table[0]

p9['Unnamed: 2'] = np.where(p9['Unnamed: 2']=='Totals', p9['Unnamed: 11'], p9['Unnamed: 2'])
p9.rename(columns = {'Unnamed: 0': 'state', 'Unnamed: 1':'agency', 'Wildland':'wildfires',
                   'Rx':'wild_acres', 'Unnamed: 2':'rx_fires',
                   'Unnamed: 3': 'rx_acres'}, inplace=True)

p9['wild_acres'].fillna(p9['Unnamed: 8'], inplace=True)
p9['wildfires'].fillna(p9['Unnamed: 5'], inplace=True)
p9['rx_acres'] = p9.rx_acres.fillna(p9['Unnamed: 14'])

p9['state'].fillna(method='ffill', inplace=True)
p9['agency'] = [i if i not in west_states else 'Totals' for i in p9['agency']]

p9.drop(p9.filter(regex='Unnamed').columns, axis=1, inplace=True)
p9.drop(index=0, inplace=True)
p9 = p9.loc[p9['state'].isin(west_states)]

p9['rx_acres'][12] = 15801
p9['rx_acres'][33] = 15197

output_path

all_2014 = pd.concat([p1_2014, p2_2014, p3, p7, p9])
all_2014.to_csv(output_path + 'annual_2014_agency.csv', index=False)
all_2014.loc[all_2014.agency == 'Totals'].to_csv(output_path + 'annual_2014_tot.csv', index=False)

## 2015

pages = [1, 2, 3, 7, 9]

pdf_list=[]

for i in pages:
    table = tabula.read_pdf(filename15, pages=i) #filename15 defined above
    pdf_list.append(table[0])

p1 = pdf_list[0].copy()

p1['Unnamed: 1'].fillna('Totals', inplace=True)
p1.drop(['Unnamed: 2', 'Unnamed: 4', 'Wildland', 'Unnamed: 6', 'Unnamed: 8', 'Rx'],
       axis=1, inplace=True)
p1.rename(columns = {'Unnamed: 0':'state', 'Unnamed: 1': 'agency', 'Unnamed: 3': 'wildfires',
                    'Unnamed: 5': 'wild_acres', 'Unnamed: 7': 'rx_fires', 'Unnamed: 9': 'rx_acres'},
         inplace=True)
p1.state.fillna(method='ffill', inplace=True)
p1 = p1.loc[p1.state.isin(west_states)]
p1

p2 = pdf_list[1].copy()

p2['Unnamed: 2'] = np.where(p2['Unnamed: 2']=='Totals', 617, p2['Unnamed: 2']) #hardcode rx_fires_tot
p2['Rx'].fillna(p2['Unnamed: 5'], inplace=True)
p2['Unnamed: 1'].fillna(p2['Unnamed: 8'], inplace=True)
p2['Unnamed: 3'] = p2['Unnamed: 3'].fillna(p2['Unnamed: 14'])
p2.rename(columns = {'Unnamed: 0': 'state', 'Wildland':'agency', 'Rx':'wildfires',
                   'Unnamed: 1':'wild_acres', 'Unnamed: 2':'rx_fires',
                   'Unnamed: 3': 'rx_acres'}, inplace=True)
p2['state'].fillna(method='ffill', inplace=True)
p2['agency'] = [i if i not in west_states else 'Totals' for i in p2['agency']]
p2.drop(p2.filter(regex='Unnamed').columns, axis=1, inplace=True)
p2.drop(index=0, inplace=True)
p2 = p2.loc[p2['state'].isin(west_states)]
p2

p3 = pdf_list[2].copy()

p3['Unnamed: 2'] = np.where(p3['Unnamed: 2']=='Totals', p3['Unnamed: 11'], p3['Unnamed: 2'])
p3.rename(columns = {'Unnamed: 0': 'state', 'Unnamed: 1':'agency', 'Wildland':'wildfires',
                   'Rx':'wild_acres', 'Unnamed: 2':'rx_fires',
                   'Unnamed: 3': 'rx_acres'}, inplace=True)
p3['wild_acres'].fillna(p3['Unnamed: 8'], inplace=True)
p3['wildfires'].fillna(p3['Unnamed: 5'], inplace=True)
p3['rx_acres'] = p3.rx_acres.fillna(p3['Unnamed: 14'])
p3['state'].fillna(method='ffill', inplace=True)
p3['agency'] = [i if i not in west_states else 'Totals' for i in p3['agency']]
p3.drop(p3.filter(regex='Unnamed').columns, axis=1, inplace=True)
p3.drop(index=0, inplace=True)
p3 = p3.loc[p3['state'].isin(west_states)] #checked against doc

p7 = pdf_list[3].copy()

p7['Unnamed: 2'] = np.where(p7['Unnamed: 2']=='Totals', p7['Unnamed: 11'], p7['Unnamed: 2'])
p7.rename(columns = {'Unnamed: 0': 'state', 'Unnamed: 1':'agency', 'Wildland':'wildfires',
                   'Rx':'wild_acres', 'Unnamed: 2':'rx_fires',
                   'Unnamed: 3': 'rx_acres'}, inplace=True)
p7['wild_acres'].fillna(p7['Unnamed: 8'], inplace=True)
p7['wildfires'].fillna(p7['Unnamed: 5'], inplace=True)
p7['rx_acres'] = p7.rx_acres.fillna(p7['Unnamed: 14'])
p7['state'].fillna(method='ffill', inplace=True)
p7['agency'] = [i if i not in west_states else 'Totals' for i in p7['agency']]
p7.drop(p7.filter(regex='Unnamed').columns, axis=1, inplace=True)
p7.drop(index=0, inplace=True)
p7 = p7.loc[p7['state'].isin(west_states)]
p7['rx_fires'][9] = 36
p7['rx_fires'][36] = 551
p7 #checked against doc

p9 = pdf_list[4].copy()

p9['Unnamed: 2'] = np.where(p9['Unnamed: 2']=='Totals', p9['Unnamed: 11'], p9['Unnamed: 2'])
p9.rename(columns = {'Unnamed: 0': 'state', 'Unnamed: 1':'agency', 'Wildland':'wildfires',
                   'Rx':'wild_acres', 'Unnamed: 2':'rx_fires',
                   'Unnamed: 3': 'rx_acres'}, inplace=True)
p9['wild_acres'].fillna(p9['Unnamed: 8'], inplace=True)
p9['wildfires'].fillna(p9['Unnamed: 5'], inplace=True)
p9['rx_acres'] = p9.rx_acres.fillna(p9['Unnamed: 14'])
p9['state'].fillna(method='ffill', inplace=True)
p9['agency'] = [i if i not in west_states else 'Totals' for i in p9['agency']]
p9.drop(p9.filter(regex='Unnamed').columns, axis=1, inplace=True)
p9.drop(index=0, inplace=True)
p9 = p9.loc[p9['state'].isin(west_states)]
p9

all_2015 = pd.concat([p1, p2, p3, p7, p9])

output_path

all_2015.reset_index(drop=True, inplace=True)
all_2015.to_csv(output_path+'annual_2015_agency.csv', index=False)
all_2015.loc[all_2015['agency']=='Totals'].to_csv(output_path+'annual_2015_tot.csv', index=False)

## Literally just type out 2016 like a plebe because it's a disaster
States split across pages
